+++ src/components/SearchForm/index.tsx
@@ -1,8 +1,8 @@
 import { ChangeEvent, FormEvent, useState } from "react";
 import { MagnifyingGlass } from "phosphor-react";
 
-import Button from "../Button";
-import Input from "../Input";
+import Button from "@/components/Button";
+import Input from "@/components/Input";
 
 import * as S from "./styles";
