+++ src/components/TransactionTypeButton/index.tsx
@@ -1,6 +1,7 @@
 import { forwardRef, ForwardRefRenderFunction } from "react";
 
 import * as S from "./styles";
+
 import { TransactionTypeButtonProps } from "./transactionTypeButton";
 
 const TransactionTypeButton: ForwardRefRenderFunction<
